platform/iot/mqtt_client/requirements.txt
paho-mqtt==1.6.1
PyJWT==2.4.0
cryptography==37.0.4
google-cloud-logging==3.2.1
google-cloud-storage==2.4.0
google-cloud-compute==1.4.0
